# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Example Golang REST API",
    "title": "Go Example REST API",
    "contact": {
      "name": "AJackTi",
      "url": "https://github.com/AJackTi",
      "email": "dtrong97vn@gmail.com"
    },
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "paths": {
...
    "/comments": {
      "post": {
        "description": "create new comment",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Comments"],
        "summary": "Create new comment",
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/models.Comment"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestErr"
            }
          }
        }
      }
    },
    "/comments/byNewsId/{id}": {
      "get": {
        "description": "Get all comment by news id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Comments"],
        "summary": "Get comments by news",
        "parameters": [
          {
            "type": "integer",
            "description": "news_id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "page",
            "description": "page number",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "size",
            "description": "number of elements per page",
            "name": "size",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "orderBy",
            "description": "filter name",
            "name": "orderBy",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.CommentsList"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestErr"
            }
          }
        }
      }
    },
    "/comments/{id}": {
      "get": {
        "description": "Get comment by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Comments"],
        "summary": "Get comment",
        "parameters": [
          {
            "type": "integer",
            "description": "comment_id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.Comment"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestErr"
            }
          }
        }
      },
      "put": {
        "description": "update new comment",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Comments"],
        "summary": "Update comment",
        "parameters": [
          {
            "type": "integer",
            "description": "comment_id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.Comment"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestErr"
            }
          }
        }
      },
      "delete": {
        "description": "delete comment",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Comments"],
        "summary": "Delete comment",
        "parameters": [
          {
            "type": "integer",
            "description": "comment_id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httpErrors.RestErr"
            }
          }
        }
      }
    },
    "/news": {
      "get": {
        "description": "Get all news with pagination",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["News"],
        "summary": "Get all news",
        "parameters": [
          {
            "type": "integer",
            "format": "page",
            "description": "page number",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "size",
            "description": "number of elements per page",
            "name": "size",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "orderBy",
            "description": "filter name",
            "name": "orderBy",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.NewsList"
            }
          }
        }
      }
    },
...
    "/news/search": {
      "get": {
        "description": "Search news by title",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["News"],
        "summary": "Search by title",
        "parameters": [
          {
            "type": "integer",
            "format": "page",
            "description": "page number",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "size",
            "description": "number of elements per page",
            "name": "size",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "orderBy",
            "description": "filter name",
            "name": "orderBy",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.NewsList"
            }
          }
        }
      }
    },
    "/news/{id}": {
      "get": {
        "description": "Get by id news handler",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["News"],
        "summary": "Get by id news",
        "parameters": [
          {
            "type": "integer",
            "description": "news_id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.News"
            }
          }
        }
      },
      "put": {
        "description": "Update news handler",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["News"],
        "summary": "Update news",
        "parameters": [
          {
            "type": "integer",
            "description": "news_id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.News"
            }
          }
        }
      },
      "delete": {
        "description": "Delete by id news handler",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["News"],
        "summary": "Delete news",
        "parameters": [
          {
            "type": "integer",
            "description": "news_id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "httpErrors.RestErr": {
      "type": "object"
    },
...
    "models.Comment": {
      "type": "object",
      "required": ["author_id", "message", "news_id"],
      "properties": {
        "author_id": {
          "type": "string"
        },
        "comment_id": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "likes": {
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "news_id": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
    "models.CommentBase": {
      "type": "object",
      "required": ["author", "author_id", "message"],
      "properties": {
        "author": {
          "type": "string"
        },
        "author_id": {
          "type": "string"
        },
        "avatar_url": {
          "type": "string"
        },
        "comment_id": {
          "type": "string"
        },
        "likes": {
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
    "models.CommentsList": {
      "type": "object",
      "properties": {
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.CommentBase"
          }
        },
        "has_more": {
          "type": "boolean"
        },
        "page": {
          "type": "integer"
        },
        "size": {
          "type": "integer"
        },
        "total_count": {
          "type": "integer"
        },
        "total_pages": {
          "type": "integer"
        }
      }
    },
    "models.News": {
      "type": "object",
      "required": ["author_id", "content", "title"],
      "properties": {
        "author_id": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "image_url": {
          "type": "string"
        },
        "news_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
    "models.NewsList": {
      "type": "object",
      "properties": {
        "has_more": {
          "type": "boolean"
        },
        "news": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.News"
          }
        },
        "page": {
          "type": "integer"
        },
        "size": {
          "type": "integer"
        },
        "total_count": {
          "type": "integer"
        },
        "total_pages": {
          "type": "integer"
        }
      }
    },
...
  }
}